[bugfix] Keep category.py compatible for Python 2 and Python 3
There is a different behaviour between py2 and py3:
5/2 gives 2 in py2 but 2.5 in py3
math.ceil gives a floating number in py2 and int in py 3

Bug: T196865

+++ scripts/category.py
@@ -1055,7 +1055,7 @@ class CategoryTidyRobot(Bot, CategoryPreprocess):
                 # can we can output in two columns?
                 count = len(cat_list)
                 if count > 1 and len(max(cat_list, key=len)) <= 31:
-                    new_column = math.ceil(count / 2)
+                    new_column = int(math.ceil(count / 2.0))
                 else:
                     new_column = 0
